Connect ai-engine & Update APILog aggregation

+++ sentryflow/types/metrics.go
@@ -9,7 +9,7 @@ import (
 // PerAPICount Structure
 type PerAPICount struct {
 	Api   string
-	Count int
+	Count uint64
 }
 
 type DbAccessLogType struct {
